backend/requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.0
python-socketio>=5.11.0
python-engineio>=4.9.0
aiohttp>=3.10.0
aiofiles>=24.1.0
aiosqlite>=0.20.0
SQLAlchemy>=2.0.30
alembic>=1.13.0
pydantic>=2.9.0
pydantic-settings>=2.5.0
sgp4>=2.23
skyfield>=1.50
apscheduler>=3.10.0
requests>=2.31.0
httpx>=0.27.0
python-dateutil>=2.9.0
pytz>=2024.1
colorlog>=6.8.0
starlette>=0.40.0
anyio>=4.4.0
websockets>=12.0
